% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{date2ymd}
\alias{date2ymd}
\title{date2ymd}
\usage{
date2ymd(dates, names = c("Year", "Month", "Day"))
}
\arguments{
\item{dates}{vector con fechas}

\item{names}{contiene 'Year','Month' and 'Day'}
}
\value{
dt.out   year, month, day
}
\description{
Transforma un vector de fechas en una lista  con year, month, day
}
