% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{ddate2date.3d}
\alias{ddate2date.3d}
\title{ddate2date.3d}
\usage{
ddate2date.3d(ddate, type = "start", with = 0.1)
}
\arguments{
\item{ddate}{A numeric value in years with decimals}

\item{type}{!no implementado!}

\item{with}{Width of the temporal granule that begins in ddate. By default 0.1 year}
}
\value{
date.table con 3 fechas: start.date, end.date, central.date
}
\description{
Transform a date in decimal year format to  range of days
}
\examples{
ddate2date.3d(1985 + seq(0,1,by=.1))
ddate2date.3d(1984 + seq(0,1,by=.1))
}
\seealso{
ddate2date
}
