% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{ddate2date}
\alias{ddate2date}
\title{ddate2date}
\usage{
ddate2date(ddate, type = "start")
}
\arguments{
\item{ddate}{A numeric value in years with decimals}

\item{type}{Not yet implemented}
}
\value{
date.table con solo una fechas: start.date
}
\description{
Transform a date in decimal year format to a specific day or a range of days
}
\examples{
ddate2date(1985 + seq(0,1,by=.1))
ddate2date(1984 + seq(0,1,by=.1))
}
