% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{Episod2ymd}
\alias{Episod2ymd}
\title{Episod2ymd}
\usage{
Episod2ymd(start.date, end.date)
}
\arguments{
\item{start.date}{Vector with start dates of episod}

\item{end.date}{Vector with ending dates of the episode}
}
\value{
data.table con columnas: Start_year	Start_month	Start_day
                                   End_year	End_month	End_day
}
\description{
Transform two dates (defining an episode) into a data.table with 6 columns:
}
\details{
Start_Year	Start_Month	Start_Day
End_Year	End_Month	End_Day
}
\examples{
    dt.date <-  ddate2date.3d(1985 + seq(0,1,by=.1))
    Episod2ymd(dt.date$start.date,dt.date$end.date)
}
