% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{granule.data2ymd}
\alias{granule.data2ymd}
\title{granule.data2ymd}
\usage{
granule.data2ymd(dt.date)
}
\arguments{
\item{dt.date}{data.table with columns: start.date end.date
central.date}
}
\value{
data.table with columns: Year Month Day Start_year
    Start_month Start_day End_year End_month End_day
}
\description{
Transform a table of 3 columns in format dates to table with 9
 columns of integers (convert each date into 3 integers: year,
 month, day)
}
\examples{
    dt.date <-  ddate2date.3d(1985 + seq(0,1,by=.1))
    granule.data2ymd(dt.date)

}
