% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ids.skeleton}
\alias{ids.skeleton}
\title{IDS tables}
\format{A list  with 4 data.tables
\describe{
  \item{INDIVIDUAL}{attributes of individuals}
  \item{CONTEXT}{attributes of the contextsos}
  \item{INDIV_INDIV}{relationships between individuals}
  \item{INDIV_CONTEX}{relationships between individuals and contexts}
  \item{CONTEX_CONTEX}{relationships between contexts}
}}
\source{
\url{https://disseminate.objectrepository.org/file/master/10622/23526343-2017-0004}
}
\usage{
ids.skeleton
}
\description{
Contains an empty structure list of the tables Intermediate Data Structure (IDS)
}
\keyword{datasets}
