% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposer.R
\name{transposer}
\alias{transposer}
\title{transposer}
\usage{
transposer(file.definition = "EntityRelationDefinition.xls",
  sheet = c("Entity", "Relationship"),
  Name.DataBase = "DB release_0.1", db.filename = "NA")
}
\arguments{
\item{file.definition}{xlxs con las especificaciones de la expresión de importanción
relativas a los objetos data.table que deben de estar previamente
presentes en .GlobalEnv}

\item{sheet}{nombre de la hoja dentro del libro}

\item{Name.DataBase}{de la BD en IDS}

\item{db.filename}{(aun no implemetado, DB SQLite donde se almacenara
la estructura IDS)}
}
\value{
outcome  una lista con las tablas INDIVIDUAL, CONTEXT, ...
         con los resultados de la carga  (en versiones porterires se alamacenara
         sobre una base de datos SQLite)
}
\description{
Funcion principal para importar un conjunto de tablas presentes
   en  .GlobalEnv sobre una estructura IDS
}
