% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{date2ymd}
\alias{date2ymd}
\title{date2ymd}
\usage{
date2ymd(dates, names = c("Year", "Month", "Day"))
}
\arguments{
\item{dates}{vector with dates}

\item{names}{Contains 'Year','Month' and 'Day'}
}
\value{
dt.out   data.table with column Year, Month and Day
}
\description{
Transform a date vector into a list with  year, month, day
}
