% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{ddate2date}
\alias{ddate2date}
\title{ddate2date}
\usage{
ddate2date(ddate, type = "start", with = 0.1)
}
\arguments{
\item{ddate}{Un valor numerico con el significado de años decimales (???).}

\item{type}{!no implementado!}

\item{with}{!no implementado!}
}
\value{
date.table con solo una fechas: start.date
}
\description{
Transforma una fecha en formato decimal a formato fecha (puntual o intervalo)
}
\examples{
ddate2date(1985 + seq(0,1,by=.1))
ddate2date(1984 + seq(0,1,by=.1))
ddate2date(1988 + seq(0,1,by=.1))
}
