% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{Episod2ymd}
\alias{Episod2ymd}
\title{Episod2ymd}
\usage{
Episod2ymd(start.date, end.date)
}
\arguments{
\item{start.date}{vector de fechas de comienzo}

\item{end.date}{vector de fechas de termininacion del episodio}
}
\value{
data.table con columnas: Start_year	Start_month	Start_day
                                   End_year	End_month	End_day
}
\description{
Transforma dos fechas(definen un episodio) en un data.tabla con 6 columnas:
Start_Year	Start_Month	Start_Day
End_Year	End_Month	End_Day
}
\examples{
    dt.date <-  ddate2date.3d(1985 + seq(0,1,by=.1))
    Episod2ymd(dt.date$start.date,dt.date$end.date)
}
