% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTranformation.R
\name{granule.data2ymd}
\alias{granule.data2ymd}
\title{granule.data2ymd}
\usage{
granule.data2ymd(dt.date)
}
\arguments{
\item{dt.date}{tabla con columnas: start.date   end.date central.date}
}
\value{
data.table con columnas: Year	Month	Day	Start_Year	Start_Month	Start_Day	End_Year	End_Month	End_Day
}
\description{
Matriz de 3 fechas a matiz númerica de 9 elementos
}
\examples{
    dt.date <-  ddate2date.3d(1985 + seq(0,1,by=.1))
    granule.data2ymd(dt.date)

}
