% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ids.skeleton}
\alias{ids.skeleton}
\title{Esquema de las tablas del IDS}
\format{A list  with 4 data.tables
\describe{
  \item{INDIVIDUAL}{tabla de atributos persona}
  \item{CONTEXT}{tabla de atributos contexto}
  \item{INDIV_INDIV}{relaciones entre individuos}
  \item{INDIV_CONTEX}{I-c}
  \item{CONTEX_CONTEX}{c-c}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
ids.skeleton
}
\description{
Contiene una lista de estructura vacias de las tablas de IDS
diamonds.
}
\keyword{datasets}
