% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposer.R
\name{transposer}
\alias{transposer}
\title{transposer}
\usage{
transposer(file.definition = "EntityRelationDefinition.xls",
  sheet = c("Entity", "Relationship"),
  Name.DataBase = "DB release_0.1", db.filename = "NA")
}
\arguments{
\item{file.definition}{xlsx-files with the instructions to perform
the transfer of the data contained in the data.table objects
present in .GlobalEnv to the tables required by the IDS format}

\item{sheet}{Name of the pages within the book that contain the
instructions. Only 'Entity' and/or 'Relationship' is allowed.}

\item{Name.DataBase}{Name of the database that identifies the data
within the IDS network.}

\item{db.filename}{Not yet implemented. The name will be the SQLite
database where the results will be stored in IDS format.}
}
\value{
outcome A list with the tables INDIVIDUAL, CONTEXT,
    INDIV_INDIV, INDIV_CONTEXT, CONTEXT_CONTEXT with the results of
    the data transfer,
}
\description{
Main function to tranfer a set of data.table(s) present in
 .GlobalEnv on an IDS structure
}
